requirements.txt
streamlit
pandas
numpy
plotly
pillow
requests
schedule
google-auth
google-auth-oauthlib
jwt
yt-dlp
instabot
stripe
openai
tensorflow
torch
transformers
spacy
textblob
python-dateutil
aiohttp
httpx
ratelimit
pyyaml
scikit-learn
scipy
user-agents
qrcode
email-validator
pyotp
bcrypt
TikTokApi
opencv-python
moviepy
Twitter
instagram_private_api
seaborn
tf-keras
psutil